refactor: 해당하는 달의 첫주, 마지막주를 계산하는 로직과 수입, 지출을 계산하는 로직 로컬 hook으로 분리
